requirements.txt
python-telegram-bot==21.*
yfinance==0.2.*
python-dotenv==1.*
pendulum==3.*
apscheduler==3.*
loguru==0.7.*
curl_cffi==0.6.*